# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "lehner.michele@example.org"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/prompt/{prompt}/generate": {
            "post": {
                "tags": [
                    "prompt generate"
                ],
                "summary": "Adds a new prompt generate",
                "description": "Store a newly created resource in storage.",
                "operationId": "75d17af9df9e089db2a9dd992ed1c2f1",
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "path",
                        "description": "prompt id example",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "1": {
                                "summary": "1",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "p1",
                        "in": "query",
                        "description": "weather value example",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "sunny": {
                                "summary": "sunny",
                                "value": "sunny"
                            },
                            "wendy": {
                                "summary": "wendy",
                                "value": "wendy"
                            }
                        }
                    },
                    {
                        "name": "p2",
                        "in": "query",
                        "description": "prompt value example",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "banging clams": {
                                "summary": "banging clams",
                                "value": "banging clams"
                            },
                            "catching fish": {
                                "summary": "catching fish",
                                "value": "catching fish"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PromptGenerateResources"
                                        }
                                    },
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Result"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Result": {
                "title": "Sample schema for using references",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "A simple of response message"
                    },
                    "data": {
                        "type": "Object"
                    }
                },
                "type": "object"
            },
            "BaseResources": {
                "properties": {
                    "id": {
                        "description": "primary key",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2023-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2023-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
            "PromptGenerateResources": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseResources"
                    },
                    {
                        "properties": {
                            "url": {
                                "description": "image url",
                                "type": "string"
                            }
                        }
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}